Finally release fabric server from the clasps of DynmapPlugin
(cherry picked from commit 339a0adb8c3e47810a5129a7c5b87ecff41f0f08)
