+++ src/main/java/org/dynmap/fabric_1_16_1/DynmapPlugin.java
@@ -1,17 +1,13 @@
 package org.dynmap.fabric_1_16_1;
 
-import com.mojang.authlib.GameProfile;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import com.sun.nio.zipfs.ZipFileSystem;
 import it.unimi.dsi.fastutil.longs.Long2ObjectLinkedOpenHashMap;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerChunkEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents;
-import net.fabricmc.loader.api.FabricLoader;
-import net.fabricmc.loader.api.ModContainer;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.FluidBlock;
@@ -19,21 +15,13 @@ import net.minecraft.block.Material;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.network.ClientConnection;
-import net.minecraft.network.MessageType;
-import net.minecraft.server.BannedIpList;
-import net.minecraft.server.BannedPlayerList;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.PlayerManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ChunkHolder;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.text.LiteralText;
-import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.UserCache;
-import net.minecraft.util.Util;
 import net.minecraft.util.collection.IdList;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.ChunkPos;
@@ -62,35 +50,26 @@ import org.dynmap.fabric_1_16_1.permissions.PermissionProvider;
 import org.dynmap.permissions.PermissionsHandler;
 import org.dynmap.renderer.DynmapBlockState;
 import org.dynmap.utils.DynmapLogger;
-import org.dynmap.utils.MapChunkCache;
-import org.dynmap.utils.VisibilityLimit;
 
 import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
 import java.util.concurrent.*;
-import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 public class DynmapPlugin {
-    private DynmapCore core;
+    // FIXME: Fix package-private fields after splitting is done
+    DynmapCore core;
     private PermissionProvider permissions;
     private boolean core_enabled;
     public SnapshotCache sscache;
     public PlayerList playerList;
-    private MapManager mapManager;
+    MapManager mapManager;
     /**
      * Server is set when running and unset at shutdown.
      */
     private net.minecraft.server.MinecraftServer server;
     public static DynmapPlugin plugin;
-    private ChatHandler chathandler;
+    ChatHandler chathandler;
     private HashMap<String, Integer> sortWeights = new HashMap<String, Integer>();
     // Drop world load ticket after 30 seconds
     private long worldIdleTimeoutNS = 30 * 1000000000L;
@@ -103,11 +82,11 @@ public class DynmapPlugin {
     private FabricServer fserver;
     private boolean tickregistered = false;
     // TPS calculator
-    private double tps;
-    private long lasttick;
-    private long avgticklen;
+    double tps;
+    long lasttick;
+    long avgticklen;
     // Per tick limit, in nsec
-    private long perTickLimit = (50000000); // 50 ms
+    long perTickLimit = (50000000); // 50 ms
     private boolean useSaveFolder = true;
 
     private static final int SIGNPOST_ID = 63;
@@ -115,7 +94,7 @@ public class DynmapPlugin {
 
     private static final String[] TRIGGER_DEFAULTS = {"blockupdate", "chunkpopulate", "chunkgenerate"};
 
-    private static final Pattern patternControlCode = Pattern.compile("(?i)\\u00A7[0-9A-FK-OR]");
+    static final Pattern patternControlCode = Pattern.compile("(?i)\\u00A7[0-9A-FK-OR]");
 
     DynmapPlugin() {
         plugin = this;
@@ -274,7 +253,7 @@ public class DynmapPlugin {
         return nh.connection;
     }
 
-    private FabricPlayer getOrAddPlayer(ServerPlayerEntity player) {
+    FabricPlayer getOrAddPlayer(ServerPlayerEntity player) {
         String name = player.getName().getString();
         FabricPlayer fp = players.get(name);
         if (fp != null) {
@@ -286,35 +265,12 @@ public class DynmapPlugin {
         return fp;
     }
 
-    private static class TaskRecord implements Comparable<Object> {
-        private long ticktorun;
-        private long id;
-        private FutureTask<?> future;
-
-        @Override
-        public int compareTo(Object o) {
-            TaskRecord tr = (TaskRecord) o;
-
-            if (this.ticktorun < tr.ticktorun) {
-                return -1;
-            } else if (this.ticktorun > tr.ticktorun) {
-                return 1;
-            } else if (this.id < tr.id) {
-                return -1;
-            } else if (this.id > tr.id) {
-                return 1;
-            } else {
-                return 0;
-            }
-        }
-    }
-
-    private static class ChatMessage {
+    static class ChatMessage {
         String message;
         ServerPlayerEntity sender;
     }
 
-    private ConcurrentLinkedQueue<ChatMessage> msgqueue = new ConcurrentLinkedQueue<ChatMessage>();
+    ConcurrentLinkedQueue<ChatMessage> msgqueue = new ConcurrentLinkedQueue<ChatMessage>();
 
     public static class ChatHandler {
         private final DynmapPlugin plugin;
@@ -399,14 +355,15 @@ public class DynmapPlugin {
 
     public boolean isOp(String player) {
         String[] ops = server.getPlayerManager().getOpList().getNames();
+
         for (String op : ops) {
             if (op.equalsIgnoreCase(player)) {
                 return true;
             }
         }
 
-        // TODO: Consider whether cheats are enabled
-        return (server.isSinglePlayer() && player.equalsIgnoreCase(server.getUserName()));
+        // TODO: Consider whether cheats are enabled for integrated server
+        return server.isSinglePlayer() && player.equalsIgnoreCase(server.getUserName());
     }
 
     boolean hasPerm(PlayerEntity psender, String permission) {
@@ -425,7 +382,7 @@ public class DynmapPlugin {
         return permissions.hasPermissionNode(psender, permission);
     }
 
-    private Set<String> hasOfflinePermissions(String player, Set<String> perms) {
+    Set<String> hasOfflinePermissions(String player, Set<String> perms) {
         Set<String> rslt = null;
         PermissionsHandler ph = PermissionsHandler.getHandler();
         if (ph != null) {
@@ -442,7 +399,7 @@ public class DynmapPlugin {
         return rslt;
     }
 
-    private boolean hasOfflinePermission(String player, String perm) {
+    boolean hasOfflinePermission(String player, String perm) {
         PermissionsHandler ph = PermissionsHandler.getHandler();
         if (ph != null) {
             if (ph.hasOfflinePermission(player, perm)) {
@@ -452,569 +409,8 @@ public class DynmapPlugin {
         return permissions.hasOfflinePermission(player, perm);
     }
 
-    /**
-     * Server access abstraction class
-     */
-    public static class FabricServer extends DynmapServerInterface {
-        /* Server thread scheduler */
-        private final Object schedlock = new Object();
-        private final DynmapPlugin plugin;
-        private final MinecraftServer server;
-        private long cur_tick;
-        private long next_id;
-        private long cur_tick_starttime;
-        private PriorityQueue<TaskRecord> runqueue = new PriorityQueue<TaskRecord>();
-
-        public FabricServer(DynmapPlugin plugin, MinecraftServer server) {
-            this.plugin = plugin;
-            this.server = server;
-        }
-
-        private GameProfile getProfileByName(String player) {
-            UserCache cache = server.getUserCache();
-            return cache.findByName(player);
-        }
-
-        @Override
-        public int getBlockIDAt(String wname, int x, int y, int z) {
-            return -1;
-        }
-
-        @Override
-        public int isSignAt(String wname, int x, int y, int z) {
-            return -1;
-        }
-
-        @Override
-        public void scheduleServerTask(Runnable run, long delay) {
-            TaskRecord tr = new TaskRecord();
-            tr.future = new FutureTask<Object>(run, null);
-
-            /* Add task record to queue */
-            synchronized (schedlock) {
-                tr.id = next_id++;
-                tr.ticktorun = cur_tick + delay;
-                runqueue.add(tr);
-            }
-        }
-
-        @Override
-        public DynmapPlayer[] getOnlinePlayers() {
-            if (server.getPlayerManager() == null) return new DynmapPlayer[0];
-
-            List<ServerPlayerEntity> players = server.getPlayerManager().getPlayerList();
-            int playerCount = players.size();
-            DynmapPlayer[] dplay = new DynmapPlayer[players.size()];
-
-            for (int i = 0; i < playerCount; i++) {
-                ServerPlayerEntity player = players.get(i);
-                dplay[i] = plugin.getOrAddPlayer(player);
-            }
-
-            return dplay;
-        }
-
-        @Override
-        public void reload() {
-            plugin.onDisable();
-            plugin.onEnable();
-            plugin.onStart();
-        }
-
-        @Override
-        public DynmapPlayer getPlayer(String name) {
-            List<ServerPlayerEntity> players = server.getPlayerManager().getPlayerList();
-
-            for (ServerPlayerEntity player : players) {
-
-                if (player.getName().getString().equalsIgnoreCase(name)) {
-                    return plugin.getOrAddPlayer(player);
-                }
-            }
-
-            return null;
-        }
-
-        @Override
-        public Set<String> getIPBans() {
-            BannedIpList bl = server.getPlayerManager().getIpBanList();
-            Set<String> ips = new HashSet<String>();
-
-            for (String s : bl.getNames()) {
-                ips.add(s);
-            }
-
-            return ips;
-        }
-
-        @Override
-        public <T> Future<T> callSyncMethod(Callable<T> task) {
-            return callSyncMethod(task, 0);
-        }
-
-        public <T> Future<T> callSyncMethod(Callable<T> task, long delay) {
-            TaskRecord tr = new TaskRecord();
-            FutureTask<T> ft = new FutureTask<T>(task);
-            tr.future = ft;
-
-            /* Add task record to queue */
-            synchronized (schedlock) {
-                tr.id = next_id++;
-                tr.ticktorun = cur_tick + delay;
-                runqueue.add(tr);
-            }
-
-            return ft;
-        }
-
-        @Override
-        public String getServerName() {
-            String sn;
-            if (server.isSinglePlayer())
-                sn = "Integrated";
-            else
-                sn = server.getServerIp();
-            if (sn == null) sn = "Unknown Server";
-            return sn;
-        }
-
-        @Override
-        public boolean isPlayerBanned(String pid) {
-            BannedPlayerList bl = server.getPlayerManager().getUserBanList();
-            return bl.contains(getProfileByName(pid));
-        }
-
-        @Override
-        public String stripChatColor(String s) {
-            return patternControlCode.matcher(s).replaceAll("");
-        }
-
-        private Set<DynmapListenerManager.EventType> registered = new HashSet<DynmapListenerManager.EventType>();
-
-        @Override
-        public boolean requestEventNotification(DynmapListenerManager.EventType type) {
-            if (registered.contains(type)) {
-                return true;
-            }
-
-            switch (type) {
-                case WORLD_LOAD:
-                case WORLD_UNLOAD:
-                    /* Already called for normal world activation/deactivation */
-                    break;
-
-                case WORLD_SPAWN_CHANGE:
-                    /*TODO
-                    pm.registerEvents(new Listener() {
-                        @EventHandler(priority=EventPriority.MONITOR)
-                        public void onSpawnChange(SpawnChangeEvent evt) {
-                            DynmapWorld w = new BukkitWorld(evt.getWorld());
-                            core.listenerManager.processWorldEvent(EventType.WORLD_SPAWN_CHANGE, w);
-                        }
-                    }, DynmapPlugin.this);
-                    */
-                    break;
-
-                case PLAYER_JOIN:
-                case PLAYER_QUIT:
-                    /* Already handled */
-                    break;
-
-                case PLAYER_BED_LEAVE:
-                    /*TODO
-                    pm.registerEvents(new Listener() {
-                        @EventHandler(priority=EventPriority.MONITOR)
-                        public void onPlayerBedLeave(PlayerBedLeaveEvent evt) {
-                            DynmapPlayer p = new BukkitPlayer(evt.getPlayer());
-                            core.listenerManager.processPlayerEvent(EventType.PLAYER_BED_LEAVE, p);
-                        }
-                    }, DynmapPlugin.this);
-                    */
-                    break;
-
-                case PLAYER_CHAT:
-                    if (plugin.chathandler == null) {
-                        plugin.chathandler = new ChatHandler(plugin);
-                        ServerChatEvents.EVENT.register((player, message) -> plugin.chathandler.handleChat(player, message));
-                    }
-                    break;
-
-                case BLOCK_BREAK:
-                    /*TODO
-                    pm.registerEvents(new Listener() {
-                        @EventHandler(priority=EventPriority.MONITOR)
-                        public void onBlockBreak(BlockBreakEvent evt) {
-                            if(evt.isCancelled()) return;
-                            Block b = evt.getBlock();
-                            if(b == null) return;
-                            Location l = b.getLocation();
-                            core.listenerManager.processBlockEvent(EventType.BLOCK_BREAK, b.getType().getId(),
-                                    BukkitWorld.normalizeWorldName(l.getWorld().getName()), l.getBlockX(), l.getBlockY(), l.getBlockZ());
-                        }
-                    }, DynmapPlugin.this);
-                    */
-                    break;
-
-                case SIGN_CHANGE:
-                    /*TODO
-                    pm.registerEvents(new Listener() {
-                        @EventHandler(priority=EventPriority.MONITOR)
-                        public void onSignChange(SignChangeEvent evt) {
-                            if(evt.isCancelled()) return;
-                            Block b = evt.getBlock();
-                            Location l = b.getLocation();
-                            String[] lines = evt.getLines();
-                            DynmapPlayer dp = null;
-                            Player p = evt.getPlayer();
-                            if(p != null) dp = new BukkitPlayer(p);
-                            core.listenerManager.processSignChangeEvent(EventType.SIGN_CHANGE, b.getType().getId(),
-                                    BukkitWorld.normalizeWorldName(l.getWorld().getName()), l.getBlockX(), l.getBlockY(), l.getBlockZ(), lines, dp);
-                        }
-                    }, DynmapPlugin.this);
-                    */
-                    break;
-
-                default:
-                    Log.severe("Unhandled event type: " + type);
-                    return false;
-            }
-
-            registered.add(type);
-            return true;
-        }
-
-        @Override
-        public boolean sendWebChatEvent(String source, String name, String msg) {
-            return DynmapCommonAPIListener.fireWebChatEvent(source, name, msg);
-        }
-
-        @Override
-        public void broadcastMessage(String msg) {
-            Text component = new LiteralText(msg);
-            server.getPlayerManager().broadcastChatMessage(component, MessageType.SYSTEM, Util.NIL_UUID);
-            Log.info(stripChatColor(msg));
-        }
-
-        @Override
-        public String[] getBiomeIDs() {
-            BiomeMap[] b = BiomeMap.values();
-            String[] bname = new String[b.length];
-
-            for (int i = 0; i < bname.length; i++) {
-                bname[i] = b[i].toString();
-            }
-
-            return bname;
-        }
-
-        @Override
-        public double getCacheHitRate() {
-            if (plugin.sscache != null)
-                return plugin.sscache.getHitRate();
-            return 0.0;
-        }
-
-        @Override
-        public void resetCacheStats() {
-            if (plugin.sscache != null)
-                plugin.sscache.resetStats();
-        }
-
-        @Override
-        public DynmapWorld getWorldByName(String wname) {
-            return plugin.getWorldByName(wname);
-        }
-
-        @Override
-        public DynmapPlayer getOfflinePlayer(String name) {
-            /*
-            OfflinePlayer op = getServer().getOfflinePlayer(name);
-            if(op != null) {
-                return new BukkitPlayer(op);
-            }
-            */
-            return null;
-        }
-
-        @Override
-        public Set<String> checkPlayerPermissions(String player, Set<String> perms) {
-            PlayerManager scm = server.getPlayerManager();
-            if (scm == null) return Collections.emptySet();
-            BannedPlayerList bl = scm.getUserBanList();
-            if (bl == null) return Collections.emptySet();
-            if (bl.contains(getProfileByName(player))) {
-                return Collections.emptySet();
-            }
-            Set<String> rslt = plugin.hasOfflinePermissions(player, perms);
-            if (rslt == null) {
-                rslt = new HashSet<String>();
-                if (plugin.isOp(player)) {
-                    rslt.addAll(perms);
-                }
-            }
-            return rslt;
-        }
-
-        @Override
-        public boolean checkPlayerPermission(String player, String perm) {
-            PlayerManager scm = server.getPlayerManager();
-            if (scm == null) return false;
-            BannedPlayerList bl = scm.getUserBanList();
-            if (bl == null) return false;
-            if (bl.contains(getProfileByName(player))) {
-                return false;
-            }
-            return plugin.hasOfflinePermission(player, perm);
-        }
-
-        /**
-         * Render processor helper - used by code running on render threads to request chunk snapshot cache from server/sync thread
-         */
-        @Override
-        public MapChunkCache createMapChunkCache(DynmapWorld w, List<DynmapChunk> chunks,
-                                                 boolean blockdata, boolean highesty, boolean biome, boolean rawbiome) {
-            FabricMapChunkCache c = (FabricMapChunkCache) w.getChunkCache(chunks);
-            if (c == null) {
-                return null;
-            }
-            if (w.visibility_limits != null) {
-                for (VisibilityLimit limit : w.visibility_limits) {
-                    c.setVisibleRange(limit);
-                }
-
-                c.setHiddenFillStyle(w.hiddenchunkstyle);
-            }
-
-            if (w.hidden_limits != null) {
-                for (VisibilityLimit limit : w.hidden_limits) {
-                    c.setHiddenRange(limit);
-                }
-
-                c.setHiddenFillStyle(w.hiddenchunkstyle);
-            }
-
-            if (!c.setChunkDataTypes(blockdata, biome, highesty, rawbiome)) {
-                Log.severe("CraftBukkit build does not support biome APIs");
-            }
-
-            if (chunks.size() == 0)     /* No chunks to get? */ {
-                c.loadChunks(0);
-                return c;
-            }
-
-            //Now handle any chunks in server thread that are already loaded (on server thread)
-            final FabricMapChunkCache cc = c;
-            Future<Boolean> f = this.callSyncMethod(new Callable<Boolean>() {
-                public Boolean call() throws Exception {
-                    // Update busy state on world
-                    FabricWorld fw = (FabricWorld) cc.getWorld();
-                    //TODO
-                    //setBusy(fw.getWorld());
-                    cc.getLoadedChunks();
-                    return true;
-                }
-            }, 0);
-            try {
-                f.get();
-            } catch (CancellationException cx) {
-                return null;
-            } catch (ExecutionException xx) {
-                Log.severe("Exception while loading chunks", xx.getCause());
-                return null;
-            } catch (Exception ix) {
-                Log.severe(ix);
-                return null;
-            }
-            if (!w.isLoaded()) {
-                return null;
-            }
-            // Now, do rest of chunk reading from calling thread
-            c.readChunks(chunks.size());
-
-            return c;
-        }
-
-        @Override
-        public int getMaxPlayers() {
-            return server.getMaxPlayerCount();
-        }
-
-        @Override
-        public int getCurrentPlayers() {
-            return server.getPlayerManager().getCurrentPlayerCount();
-        }
-
-        public void tickEvent(MinecraftServer server) {
-            cur_tick_starttime = System.nanoTime();
-            long elapsed = cur_tick_starttime - plugin.lasttick;
-            plugin.lasttick = cur_tick_starttime;
-            plugin.avgticklen = ((plugin.avgticklen * 99) / 100) + (elapsed / 100);
-            plugin.tps = (double) 1E9 / (double) plugin.avgticklen;
-            // Tick core
-            if (plugin.core != null) {
-                plugin.core.serverTick(plugin.tps);
-            }
-
-            boolean done = false;
-            TaskRecord tr = null;
-
-            while (!plugin.blockupdatequeue.isEmpty()) {
-                BlockUpdateRec r = plugin.blockupdatequeue.remove();
-                BlockState bs = r.w.getBlockState(new BlockPos(r.x, r.y, r.z));
-                int idx = Block.STATE_IDS.getId(bs);
-                if (!org.dynmap.hdmap.HDBlockModels.isChangeIgnoredBlock(stateByID[idx])) {
-                    if (plugin.onblockchange_with_id)
-                        plugin.mapManager.touch(r.wid, r.x, r.y, r.z, "blockchange[" + idx + "]");
-                    else
-                        plugin.mapManager.touch(r.wid, r.x, r.y, r.z, "blockchange");
-                }
-            }
-
-            long now;
-
-            synchronized (schedlock) {
-                cur_tick++;
-                now = System.nanoTime();
-                tr = runqueue.peek();
-                /* Nothing due to run */
-                if ((tr == null) || (tr.ticktorun > cur_tick) || ((now - cur_tick_starttime) > plugin.perTickLimit)) {
-                    done = true;
-                } else {
-                    tr = runqueue.poll();
-                }
-            }
-            while (!done) {
-                tr.future.run();
-
-                synchronized (schedlock) {
-                    tr = runqueue.peek();
-                    now = System.nanoTime();
-                    /* Nothing due to run */
-                    if ((tr == null) || (tr.ticktorun > cur_tick) || ((now - cur_tick_starttime) > plugin.perTickLimit)) {
-                        done = true;
-                    } else {
-                        tr = runqueue.poll();
-                    }
-                }
-            }
-            while (!plugin.msgqueue.isEmpty()) {
-                ChatMessage cm = plugin.msgqueue.poll();
-                DynmapPlayer dp = null;
-                if (cm.sender != null)
-                    dp = plugin.getOrAddPlayer(cm.sender);
-                else
-                    dp = new FabricPlayer(plugin, null);
-
-                plugin.core.listenerManager.processChatEvent(DynmapListenerManager.EventType.PLAYER_CHAT, dp, cm.message);
-            }
-            // Check for generated chunks
-            if ((cur_tick % 20) == 0) {
-            }
-        }
-
-        private <T> Predicate<T> distinctByKeyAndNonNull(Function<? super T, ?> keyExtractor) {
-            Set<Object> seen = ConcurrentHashMap.newKeySet();
-            return t -> t != null && seen.add(keyExtractor.apply(t));
-        }
-
-        private Optional<ModContainer> getModContainerById(String id) {
-            return FabricLoader.getInstance().getModContainer(id);
-        }
-
-        @Override
-        public boolean isModLoaded(String name) {
-            return FabricLoader.getInstance().getModContainer(name).isPresent();
-        }
-
-        @Override
-        public String getModVersion(String name) {
-            Optional<ModContainer> mod = getModContainerById(name);    // Try case sensitive lookup
-            return mod.map(modContainer -> modContainer.getMetadata().getVersion().getFriendlyString()).orElse(null);
-        }
-
-        @Override
-        public double getServerTPS() {
-            return plugin.tps;
-        }
-
-        @Override
-        public String getServerIP() {
-            if (server.isSinglePlayer())
-                return "0.0.0.0";
-            else
-                return server.getServerIp();
-        }
-
-        @Override
-        public File getModContainerFile(String name) {
-            Optional<ModContainer> container = getModContainerById(name);    // Try case sensitive lookup
-            if (container.isPresent()) {
-                Path path = container.get().getRootPath();
-                if (path.getFileSystem() instanceof ZipFileSystem) {
-                    path = Paths.get(path.getFileSystem().toString());
-                }
-                return path.toFile();
-            }
-            return null;
-        }
-
-        @Override
-        public List<String> getModList() {
-            return FabricLoader.getInstance()
-                    .getAllMods()
-                    .stream()
-                    .map(container -> container.getMetadata().getId())
-                    .collect(Collectors.toList());
-        }
-
-        @Override
-        public Map<Integer, String> getBlockIDMap() {
-            Map<Integer, String> map = new HashMap<Integer, String>();
-            return map;
-        }
-
-        @Override
-        public InputStream openResource(String modid, String rname) {
-            if (modid == null) modid = "minecraft";
-
-            if ("minecraft".equals(modid)) {
-                return MinecraftServer.class.getClassLoader().getResourceAsStream(rname);
-            } else {
-                if (rname.startsWith("/") || rname.startsWith("\\")) {
-                    rname = rname.substring(1);
-                }
-
-                final String finalModid = modid;
-                final String finalRname = rname;
-                return getModContainerById(modid).map(container -> {
-                    try {
-                        return Files.newInputStream(container.getPath(finalRname));
-                    } catch (IOException e) {
-                        Log.severe("Failed to load resource of mod :" + finalModid, e);
-                        return null;
-                    }
-                }).orElse(null);
-            }
-        }
-
-        /**
-         * Get block unique ID map (module:blockid)
-         */
-        @Override
-        public Map<String, Integer> getBlockUniqueIDMap() {
-            HashMap<String, Integer> map = new HashMap<String, Integer>();
-            return map;
-        }
-
-        /**
-         * Get item unique ID map (module:itemid)
-         */
-        @Override
-        public Map<String, Integer> getItemUniqueIDMap() {
-            HashMap<String, Integer> map = new HashMap<String, Integer>();
-            return map;
-        }
-
+    void setChatHandler(ChatHandler chatHandler) {
+        plugin.chathandler = chatHandler;
     }
 
     public class TexturesPayload {
@@ -1028,47 +424,6 @@ public class DynmapPlugin {
 
     public class ProfileTexture {
         public String url;
-    }
-
-    /* Handler for generic console command sender */
-    public static class FabricCommandSender implements DynmapCommandSender {
-        private ServerCommandSource sender;
-
-        protected FabricCommandSender() {
-            sender = null;
-        }
-
-        public FabricCommandSender(ServerCommandSource send) {
-            sender = send;
-        }
-
-        @Override
-        public boolean hasPrivilege(String privid) {
-            return true;
-        }
-
-        @Override
-        public void sendMessage(String msg) {
-            if (sender != null) {
-                Text ichatcomponent = new LiteralText(msg);
-                sender.sendFeedback(ichatcomponent, false);
-            }
-        }
-
-        @Override
-        public boolean isConnected() {
-            return false;
-        }
-
-        @Override
-        public boolean isOp() {
-            return true;
-        }
-
-        @Override
-        public boolean hasPermissionNode(String node) {
-            return true;
-        }
     }
 
     public void loadExtraBiomes(String mcver) {
@@ -1250,7 +605,7 @@ public class DynmapPlugin {
         saveWorlds();
 
         /* Purge tick queue */
-        fserver.runqueue.clear();
+        fserver.clearTaskQueue();
 
         /* Disable core */
         core.disableCore();
@@ -1453,7 +808,7 @@ public class DynmapPlugin {
     private boolean onblockchange = false;
     private boolean onchunkpopulate = false;
     private boolean onchunkgenerate = false;
-    private boolean onblockchange_with_id = false;
+    boolean onblockchange_with_id = false;
 
     private void registerEvents() {
         // To trigger rendering.
@@ -1498,7 +853,7 @@ public class DynmapPlugin {
         }
     }
 
-    private FabricWorld getWorldByName(String name) {
+    FabricWorld getWorldByName(String name) {
         return worlds.get(name);
     }
 
@@ -1594,10 +949,6 @@ public class DynmapPlugin {
                 return;
             }
         }
-    }
-
-    public MinecraftServer getMCServer() {
-        return server;
     }
 }
